Renamed files and moved ids to proper util
