+++ src/utils/client-dup.ts
@@ -1,7 +1,6 @@
 import { google } from "googleapis";
 import { Client } from "../models/client";
-import { serialize_rows, sheet_auth } from "../utils/sheets";
-import { CLIENT_SHEET_ID } from "./sheetID";
+import { CLIENT_SHEET_ID, serialize_rows, sheet_auth } from "./sheets";
 
 export const check_dup_client = async (mac_address, alias) => {
     const sheet = google.sheets("v4");
