+++ src/utils/sheets.ts
@@ -1,5 +1,9 @@
 import { google } from "googleapis";
 
+export const EVENT_SHEET_ID = "EVENTS";
+export const CHECKIN_SHEET_ID = "CHECKINS";
+export const CLIENT_SHEET_ID = "CLIENT";
+
 export const sheet_auth = () => {
   let split_key = process.env.SERVICE_KEY?.split('\\n').join("\n")
